refactor(inbox): split grouping.rs into cohesive submodules
Pure structural split, no behaviour change. grouping.rs becomes the
module root (metadata, dimension, config, result, engine submodules,
engine/tests.rs); public API and re-exports unchanged.
